Update to calculate the simple interest on a given principal amount, rate of interest, and time.

+++ python_introduction/simple_interest.py
@@ -5,7 +5,7 @@ rate = 0.05
 time = 3
 
 # Formula for calculating the simple interest
-interest = float(principal * rate * time)
+interest = principal * rate * time
 
 # Print the result
-print("The simple interest is: " + str(interest))
+print("The simple interest is: " + str(float(interest)))
